Implement binary priority level for the queue
You can now queue URLs with priority 0 or 1. If the priority
is 0, then the URL is treated as a normal URL, and gets queued. If
the priority not 0, then it is treated as a priority URL and gets
processed much faster, usually instantly.

This is implemented by keeping track of two separate queues, meaning
that if you send 100 priority URLs, they will still be queued, but
that queue will start from 0 wait time and the priority requests and
non-priority requests will overlap. However, on average, the interval
should be respected.